Revert "refactor: 새 서비스 생성"
This reverts commit bb3b800f11cf94a5e6412a2c76d22c0214c4201e.
