+++ src/main/java/roomescape/presentation/ReservationController.java
@@ -17,7 +17,6 @@ import roomescape.dto.request.PaymentRequest;
 import roomescape.dto.request.ReservationCreateRequest;
 import roomescape.dto.response.ReservationResponse;
 import roomescape.dto.response.ReservationWithStatusResponse;
-import roomescape.service.PaymentService;
 import roomescape.service.ReservationService;
 
 @RestController
@@ -25,11 +24,12 @@ import roomescape.service.ReservationService;
 public class ReservationController {
 
     private final ReservationService reservationService;
-    private final PaymentService paymentService;
+    private final PaymentClientController paymentClientController;
 
-    public ReservationController(final ReservationService reservationService, final PaymentService paymentService) {
+    public ReservationController(final ReservationService reservationService,
+                                 final PaymentClientController paymentClientController) {
         this.reservationService = reservationService;
-        this.paymentService = paymentService;
+        this.paymentClientController = paymentClientController;
     }
 
     @PostMapping
@@ -37,7 +37,7 @@ public class ReservationController {
             @Authenticated Long memberId,
             @Valid @RequestBody ReservationCreateRequest request) {
         PaymentRequest paymentRequest = new PaymentRequest(request.amount(), request.paymentKey(), request.orderId());
-        PaymentInfo paymentInfo = paymentService.createPaymentInfo(paymentRequest);
+        PaymentInfo paymentInfo = paymentClientController.postPaymentInfo(paymentRequest);
 
         ReservationResponse reservationResponse = reservationService.createReservationForMember(
                 memberId, request.timeId(), request.themeId(), request.date(), paymentInfo
